backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=0.2.0
requests>=2.31.0
biopython>=1.81
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

